+++ src/vlog/common/edb.cpp
@@ -225,7 +225,7 @@ void EDBLayer::query(QSQQuery *query, TupleTable *outputTable,
                                 row[0] = itr->first;
                                 row[1] = itr->second;
                                 if (nRepeatedVars > 0) {
-                                    for (uint8_t i = 0; i < nRepeatedVars; ++i) {
+                                    for (int i = 0; i < nRepeatedVars; ++i) {
                                         std::pair<uint8_t, uint8_t> rp = query->getRepeatedVar(i);
                                         if (row[rp.first] != row[rp.second]) {
                                             valid = false;
@@ -281,7 +281,7 @@ void EDBLayer::query(QSQQuery *query, TupleTable *outputTable,
                                             row[0] = itr1->first;
                                             row[1] = itr1->second;
                                             if (nRepeatedVars > 0) {
-                                                for (uint8_t i = 0; i < nRepeatedVars; ++i) {
+                                                for (int i = 0; i < nRepeatedVars; ++i) {
                                                     std::pair<uint8_t, uint8_t> rp = query->getRepeatedVar(i);
                                                     if (row[rp.first] != row[rp.second]) {
                                                         valid = false;
@@ -317,7 +317,7 @@ void EDBLayer::query(QSQQuery *query, TupleTable *outputTable,
                                             row[0] = itr1->first;
                                             row[1] = itr1->second;
                                             if (nRepeatedVars > 0) {
-                                                for (uint8_t i = 0; i < nRepeatedVars; ++i) {
+                                                for (int i = 0; i < nRepeatedVars; ++i) {
                                                     std::pair<uint8_t, uint8_t> rp = query->getRepeatedVar(i);
                                                     if (row[rp.first] != row[rp.second]) {
                                                         valid = false;
@@ -379,7 +379,7 @@ void EDBLayer::query(QSQQuery *query, TupleTable *outputTable,
                                     row[0] = itr->first;
                                     row[1] = itr->second;
                                     if (nRepeatedVars > 0) {
-                                        for (uint8_t i = 0; i < nRepeatedVars; ++i) {
+                                        for (int i = 0; i < nRepeatedVars; ++i) {
                                             std::pair<uint8_t, uint8_t> rp = query->getRepeatedVar(i);
                                             if (row[rp.first] != row[rp.second]) {
                                                 valid = false;
@@ -616,7 +616,7 @@ bool EDBLayer::isEmpty(const Literal &query, std::vector<uint8_t> *posToFilter,
             bool foundConstant = false;
             uint8_t idxVar = 0;
             Term_t valConst = 0;
-            for (uint8_t i = 0; i < literal->getTupleSize(); ++i) {
+            for (int i = 0; i < literal->getTupleSize(); ++i) {
                 if (!literal->getTermAtPos(i).isVariable()) {
                     idxVar = i;
                     valConst = literal->getTermAtPos(i).getValue();
